# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Property Lo API — Nigerian real estate listing platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /favorites:
    get:
      operationId: getFavorites
      tags: [favorites]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Favorite listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
    post:
      operationId: addFavorite
      tags: [favorites]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteInput"
      responses:
        "200":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /favorites/{listingId}:
    delete:
      operationId: removeFavorite
      tags: [favorites]
      security:
        - bearerAuth: []
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Chats ─────────────────────────────────────────────────────────────────
  /chats:
    get:
      operationId: getChats
      tags: [chats]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User chats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chat"
    post:
      operationId: createChat
      tags: [chats]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChatInput"
      responses:
        "201":
          description: Chat ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatIdResponse"
  /chats/unread-count:
    get:
      operationId: getChatUnreadCount
      tags: [chats]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Unread count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnreadCountResponse"
  /chats/{id}/messages:
    get:
      operationId: getChatMessages
      tags: [chats]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chats]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"

  # ── Support ───────────────────────────────────────────────────────────────
  /support/chat:
    get:
      operationId: getSupportChat
      tags: [support]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Support chat messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportChatResponse"
...
  /admin/chats:
    get:
      operationId: getAdminChats
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: label
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: string
      responses:
        "200":
          description: All chats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminChat"
  /admin/chats/{id}/messages:
    get:
      operationId: getAdminChatMessages
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendAdminChatMessage
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
...
  /admin/support/chats:
    get:
      operationId: getAdminSupportChats
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Support chats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminSupportChat"
  /admin/support/chats/{chatId}/messages:
    get:
      operationId: getAdminSupportMessages
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Support messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportMessage"
    post:
      operationId: sendAdminSupportMessage
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportMessage"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    UnreadCountResponse:
      type: object
      required: [count]
      properties:
        count:
          type: integer
    ChatIdResponse:
      type: object
      required: [id]
      properties:
        id:
          type: string

    # ── Auth schemas ──────────────────────────────────────────────────────────
...
    Listing:
      type: object
      required: [id, title, description, type, purpose, price, location, state, images, status, created_at]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          type: string
        purpose:
          type: string
        price:
          type: number
        location:
          type: string
        state:
          type: string
        size:
          type: ["string", "null"]
        bedrooms:
          type: ["integer", "null"]
        bathrooms:
          type: ["integer", "null"]
        images:
          type: array
          items:
            type: string
        status:
          type: string
        contact_email:
          type: ["string", "null"]
        contact_phone:
          type: ["string", "null"]
        user_id:
          type: ["string", "null"]
        created_at:
          type: string
...
    FavoriteInput:
      type: object
      required: [listing_id]
      properties:
        listing_id:
          type: string

    # ── Chats ─────────────────────────────────────────────────────────────────
    Chat:
      type: object
      required: [id, listing_id, listing_title, label, user_unread, admin_unread, updated_at, created_at]
      properties:
        id:
          type: string
        listing_id:
          type: string
        listing_title:
          type: string
        label:
          type: string
        last_message:
          type: ["string", "null"]
        user_unread:
          type: integer
        admin_unread:
          type: integer
        updated_at:
          type: string
        created_at:
          type: string
    ChatMessage:
      type: object
      required: [id, content, is_admin, created_at]
      properties:
        id:
          type: string
        content:
          type: string
        is_admin:
          type: boolean
        created_at:
          type: string
    CreateChatInput:
      type: object
      required: [listing_id]
      properties:
        listing_id:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string

    # ── Support ───────────────────────────────────────────────────────────────
    SupportMessage:
      type: object
      required: [id, content, is_from_user, created_at]
      properties:
        id:
          type: string
        content:
          type: string
        is_from_user:
          type: boolean
        created_at:
          type: string
    SupportChatResponse:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/SupportMessage"

    # ── Ads ───────────────────────────────────────────────────────────────────
...
    AdminChat:
      type: object
      required: [id, listing_id, listing_title, label, user_unread, admin_unread, updated_at]
      properties:
        id:
          type: string
        listing_id:
          type: string
        listing_title:
          type: string
        label:
          type: string
        last_message:
          type: ["string", "null"]
        user_unread:
          type: integer
        admin_unread:
          type: integer
        updated_at:
          type: string
        user_name:
          type: ["string", "null"]
    AdminSupportChat:
      type: object
      required: [id, user_unread, admin_unread, updated_at]
      properties:
        id:
          type: string
        user_name:
          type: ["string", "null"]
        last_message:
          type: ["string", "null"]
        user_unread:
          type: integer
        admin_unread:
          type: integer
        updated_at:
          type: string
...