# lib/api-zod/src/generated/types/supportMessage.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Property Lo API — Nigerian real estate listing platform
 * OpenAPI spec version: 0.1.0
 */

export interface SupportMessage {
  id: string;
  content: string;
  is_from_user: boolean;
  created_at: string;
}